ref namespaces: support remote repositories via upload-pack and receive-pack
Teach upload-pack and receive-pack to optionally use the namespace
prefixed refs when working with the repository, and use the
unprefixed refs when talking to the client, to give an illusion that
only the refs in the namespace exists in the repository. This allows
clone, pull, fetch, and push to work with a suitably configured
GIT_NAMESPACE.

With appropriate configuration, this also allows http-backend to expose
namespaces as multiple repositories with different paths.  This only
requires setting GIT_NAMESPACE, which http-backend passes through to
upload-pack and receive-pack.

Signed-off-by: Josh Triplett <josh@joshtriplett.org>
Signed-off-by: Jamey Sharp <jamey@minilop.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>